LayoutTests/fast/viewport/viewport-warnings-1.html

<head>
    <title>No warnings or errors.</title>
    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=2">
    <script>
        function test() {
            if (window.testRunner) {
                testRunner.dumpAsText();
                alert(internals.configurationForViewport(document, 1, 320, 480, 320, 352));
            }
        }
    </script>
</head>
<body onload="test();">
