safari-600.7-branch/LayoutTests/fast/viewport/viewport-warnings-1-expected.txt

ALERT: viewport size 320x352 scale 1 with limits [1, 2] and userScalable true
